Refactor : search components into dedicated directory
Moved search-related components (SearchBar, SearchDropdown, DateStep, DestinationStep, GuestStep) from the layout directory to a new search directory for better organization. Updated imports in Header to reflect the new structure and removed inline comments for cleaner code.
