+++ frontend/components/layout/Container.tsx
@@ -1,4 +1,4 @@
-import type { ReactNode } from "react";
+import type { ReactNode } from 'react';
 
 export default function Container({ children }: { children?: ReactNode }) {
   return <div className="container mx-auto px-4">{children}</div>;
